remove executor file and made changes accordingly and updated test cases
