+++ kairon/shared/log_system/handlers/actions_logs_handler.py
@@ -6,9 +6,8 @@ from kairon.shared.log_system.base import BaseLogHandler
 
 class ActionLogHandler(BaseLogHandler):
     def get_logs_and_count(self):
-        from kairon.shared.log_system.executor import LogExecutor
         query = {"bot": self.bot, "trigger_info.trigger_id": ""}
         logs_cursor = self.doc_type.objects(**query).order_by("-timestamp").skip(self.start_idx).limit(self.page_size).exclude("bot")
         logs = json.loads(logs_cursor.to_json())
-        count = LogExecutor.get_logs_count(self.doc_type, **query)
+        count = self.get_logs_count(self.doc_type, **query)
         return logs, count
