+++ kairon/shared/log_system/handlers/callback_logs_handler.py
@@ -5,7 +5,6 @@ from kairon.shared.log_system.base import BaseLogHandler
 
 class CallbackLogHandler(BaseLogHandler):
     def get_logs_and_count(self):
-        from kairon.shared.log_system.executor import LogExecutor
         query = {"bot": self.bot}
         for field in ["name", "sender_id", "channel", "identifier"]:
             value = self.kwargs.get(field)
@@ -13,5 +12,5 @@ class CallbackLogHandler(BaseLogHandler):
                 query["callback_name" if field == "name" else field] = value
         logs_cursor = self.doc_type.objects(**query).order_by("-timestamp").skip(self.start_idx).limit(self.page_size).exclude("id", "bot", "user", "_id")
         logs = json.loads(logs_cursor.to_json())
-        count = LogExecutor.get_logs_count(self.doc_type, **query)
+        count = self.get_logs_count(self.doc_type, **query)
         return logs, count
