+++ kairon/shared/log_system/handlers/executor_logs_handler.py
@@ -5,7 +5,6 @@ from kairon.shared.log_system.base import BaseLogHandler
 
 class ExecutorLogHandler(BaseLogHandler):
     def get_logs_and_count(self):
-        from kairon.shared.log_system.executor import LogExecutor
         query = {"bot": self.bot}
         for field in ["event_class", "task_type"]:
             value = self.kwargs.get(field)
@@ -13,5 +12,5 @@ class ExecutorLogHandler(BaseLogHandler):
                 query[field] = value
         logs_cursor = self.doc_type.objects(**query).order_by("-timestamp").skip(self.start_idx).limit(self.page_size).exclude("_id")
         logs = json.loads(logs_cursor.to_json())
-        count = LogExecutor.get_logs_count(self.doc_type, **query)
+        count = self.get_logs_count(self.doc_type, **query)
         return logs, count
